% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_table.R
\name{cross_table}
\alias{cross_table}
\title{Krysstabell med replikasjonsvekter}
\usage{
cross_table(
  data,
  svy,
  row_var,
  col_var,
  prosent = c("nei", "rad", "kolonne", "total"),
  test = TRUE,
  as_gt = FALSE,
  plot = FALSE,
  return_data = FALSE,
  row_label = NULL,
  col_label = NULL
)
}
\arguments{
\item{data}{Datasettet som skal brukes}

\item{svy}{Survey-type: "TALISEC_LEADER" eller "TALISEC_STAFF"}

\item{row_var}{Raden i tabellen (tekst)}

\item{col_var}{Kolonnen i tabellen (tekst)}

\item{prosent}{Type prosent: "nei", "rad", "kolonne", "total" (default: "nei")}

\item{test}{Inkluder Rao-Scott-justert chi²-test? (default: TRUE)}

\item{as_gt}{Returner som gt-tabell? (default: FALSE)}

\item{plot}{Returner som ggplot? (default: FALSE)}

\item{return_data}{Returner data og ev. testresultat? (TRUE eller FALSE, default: FALSE)}

\item{row_label}{Egendefinert etikett for rader (valgfritt)}

\item{col_label}{Egendefinert etikett for kolonner (valgfritt)}
}
\value{
Tabell, plot eller tibble
}
\description{
Lager en krysstabell (prosent eller antall) med opsjon for test, tabell og plott.
}
