% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{Frekvenstabell med replikasjonsvekter}
\usage{
freq_table(
  data,
  svy,
  variabel,
  by = NULL,
  fast = FALSE,
  as_gt = FALSE,
  plot = FALSE,
  return_data = c("none", "tabell", "rrepest"),
  sort = FALSE,
  ci = FALSE,
  var_label = NULL,
  by_label = NULL
)
}
\arguments{
\item{data}{Datasettet som skal brukes}

\item{svy}{Survey-type: "TALISEC_LEADER" eller "TALISEC_STAFF"}

\item{variabel}{Variabelnavn som skal summeres (som tekst, f.eks. "gender")}

\item{by}{Grupperingsvariabel (valgfritt, tekst)}

\item{fast}{Bruk rask metode med færre vekter? (default: FALSE)}

\item{as_gt}{Returner som gt-tabell? (default: FALSE)}

\item{plot}{Returner som ggplot? (default: FALSE)}

\item{return_data}{Hva slags data skal returneres? "none" (default), "tabell" for formattert resultat, eller "rrepest" for årsresultatene fra Rrepest.}

\item{sort}{Sorter etter andel? (default: FALSE)}

\item{ci}{Inkluder konfidensintervall (95 \%) i resultatet? (default: FALSE)}

\item{var_label}{Navn som skal brukes i tabell eller plott i stedet for variabelnavn (valgfritt)}

\item{by_label}{Navn som skal brukes i stedet for by-variabelnavn (valgfritt)}
}
\value{
Tabell, plot, tibble eller Rrepest-resultat
}
\description{
Lager en frekvenstabell med usikkerhet, med opsjon for tabell, plot eller rådata.
}
