% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_talis_design.R
\name{get_talis_design}
\alias{get_talis_design}
\title{Returnerer survey design for TALIS Starting Strong}
\usage{
get_talis_design(data, svy)
}
\arguments{
\item{data}{Datasett som inneholder riktige vektvariabler og replikasjonsvekter}

\item{svy}{Survey-type: "TALISEC_STAFF" eller "TALISEC_LEADER"}
}
\value{
Et objekt av klassen svyrep.design
}
\description{
Lager et survey design-objekt med replikasjonsvekter for bruk med survey-pakken.
}
