% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_table.R
\name{mean_table}
\alias{mean_table}
\title{Gjennomsnittstabell med replikasjonsvekter}
\usage{
mean_table(
  data,
  svy,
  variabel,
  by = NULL,
  fast = FALSE,
  as_gt = FALSE,
  plot = FALSE,
  return_data = c("none", "tabell", "rrepest"),
  sort = FALSE,
  var_label = NULL,
  by_label = NULL,
  ci = FALSE
)
}
\arguments{
\item{data}{Datasettet som skal brukes}

\item{svy}{Survey-type: "TALISEC_LEADER" eller "TALISEC_STAFF"}

\item{variabel}{Variabelnavn (tekst, f.eks. "ss2g02")}

\item{by}{Grupperingsvariabel (valgfritt, tekst)}

\item{fast}{Bruk rask metode med færre vekter? (default: FALSE)}

\item{as_gt}{Returner som gt-tabell? (default: FALSE)}

\item{plot}{Returner som ggplot? (default: FALSE)}

\item{return_data}{Hva skal returneres? "none" (default), "tabell" eller "rrepest"}

\item{sort}{Sorter etter gjennomsnitt? (default: FALSE)}

\item{var_label}{Egendefinert etikett for variabelnavnet}

\item{by_label}{Egendefinert etikett for grupperingsvariabelen}

\item{ci}{Inkluder 95 \% konfidensintervall? (default: FALSE)}
}
\value{
Tabell, plott, eller tibble
}
\description{
Beregner gjennomsnitt og usikkerhet for en variabel, med støtte for grupper, plot, og konfidensintervall.
}
