% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_mean.R
\name{pairwise_mean}
\alias{pairwise_mean}
\title{Parvise sammenlikninger av gjennomsnitt (survey design)}
\usage{
pairwise_mean(data, svy, outcome, group, p_adjust = "holm", as_gt = FALSE)
}
\arguments{
\item{data}{Datasettet som skal brukes}

\item{svy}{Survey-type: "TALISEC_LEADER" eller "TALISEC_STAFF"}

\item{outcome}{Den kontinuerlige variabelen vi sammenlikner (tekst)}

\item{group}{Gruppene vi sammenlikner på (faktorvariabel, tekst)}

\item{p_adjust}{Metode for p-justering (default: "holm")}

\item{as_gt}{Returner som gt-tabell? (default: FALSE)

Gyldige p_adjust-metoder for funksjoner som sammenlikner gjennomsnitt:
\itemize{
\item "bonferroni" (strengest)
\item "holm" (anbefalt for parvise tester)
\item "BH" (Benjamini-Hochberg, FDR)
\item "fdr" (synonym for "BH")
\item "none" (ingen justering)
}}
}
\value{
Tabell med estimerte forskjeller
}
\description{
Estimerer gjennomsnitt for en kontinuerlig variabel innenfor grupper, og tester parvise forskjeller
}
