% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_battery.R
\name{freq_battery}
\alias{freq_battery}
\title{Frekvenstabell for spørsmålsbatterier}
\usage{
freq_battery(
  data,
  svy,
  battery_prefix,
  fast = FALSE,
  as_gt = FALSE,
  plot = FALSE,
  return_data = c("none", "tabell", "rrepest")
)
}
\arguments{
\item{data}{Datasettet som skal brukes}

\item{svy}{Survey-type: "TALISEC_LEADER" eller "TALISEC_STAFF"}

\item{battery_prefix}{Prefiks for variablene i batteriet (f.eks. "q_29")}

\item{fast}{Bruk rask metode med færre vekter? (default: FALSE)}

\item{as_gt}{Returner som gt-tabell? (default: FALSE)}

\item{plot}{Returner som ggplot? (default: FALSE)}

\item{return_data}{Hva skal returneres? "none" (default), "tabell" eller "rrepest"}
}
\value{
Tabell, plot, tibble eller liste med resultater per variabel
}
\description{
Lager en frekvenstabell for alle variabler som starter med et gitt prefix.
}
